requirements.txt
Keras==2.4.3
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.1.0
scikit-learn==0.24.1
scipy==1.1.0
sklearn==0.0
tensorflow==2.4.1
torch==1.5.0
torchvision==0.6.0
tqdm==4.56.0
scikit-image==0.14.3
openslide-python
dicom